+++ source/builtins/exit.c
@@ -6,27 +6,27 @@
 /*   By: gomandam <gomandam@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/24 02:03:44 by gomandam          #+#    #+#             */
-/*   Updated: 2025/09/11 02:31:36 by gomandam         ###   ########.fr       */
+/*   Updated: 2025/09/11 22:01:34 by gomandam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/minishell.h"
+#include "../../libft/libft.h"
 
-// TEMPORARY free_shell to be declared on minishel.h after being uploaded in branch
+// internal free
 static void	free_shell(t_shell *shell)
 {
 	if (!shell)
-		return;
+		return ;
 	if (shell->line)
 		free(shell->line);
-	if (shell->tokens)
-		free_tokens(&shell->tokens);
 	if (shell->ast)
-		free_ast_parse(&shell->ast);
+		free_exp_ast(&shell->ast);
 	free_env_list(&shell->env_list);
 }
 
-//return 1 if string is valid, 0 otherwise.
+/*	Check string = valid numeric argument for exit
+	Accepts optional + or - sign, then digits only	*/
 static int	is_numeric(const char *str)
 {
 	int	i;
@@ -34,7 +34,7 @@ static int	is_numeric(const char *str)
 	i = 0;
 	if (!str || !str[0])
 		return (0);
-	if (str[i] == '-' || str[i] == '+')
+	if (str[0] == '-' || str[0] == '+')
 		i++;
 	while (str[i])
 	{
@@ -44,32 +44,70 @@ static int	is_numeric(const char *str)
 	}
 	return (1);
 }
+//! 	string to long long (ft_strtoll) => check this functions
+//	Converts string to long long
 
-int	ft_exit(t_shell *shell, char *argv[])
+static long long	ft_strtoll(const char *str)
 {
-	int	exit_status;
+	long long	res;
+	int			sign;
+	int			i;
 
-	exit_status = 0;
-	write(1, "exit\n", 5);
-	if (argv[1])
+	res = 0;
+	sign = 1;
+	i = 0;
+	while (str[i] == ' ' || (str[i] >= 9 && str[i] <= 13))
+		i++;
+	if (str[i] == '-' || str[i] == '+')
 	{
-		if (!is_numeric(argv[1]))
-		{
-			printf("minishell: exit: %s: numeric argument required\n", argv[1]);
-			exit_status = 255;
-		}
-		else if (argv[2])
-		{
-			printf("exit: too many arguments\n");
-			return (1);
-		}
-		else
-			exit_status = (unsigned char)ft_atoi(argv[1]);	// unsigned char to ensure UNIX exit codes [0, 255]
+		if (str[i] == '-')
+			sign = -1;
+		i++;
 	}
-	rl_clear_history();		// Free the readline history memory, since exit
-	free_shell(shell);		// ! TO DO IMPLEMENT: Free all resources and structures
-	exit(exit_status);
+	while (str[i] >= '0' && str[i] <= '9')
+	{
+		res = res * 10 + (str[i] - '0');
+		i++;
+	}
+	return (res * sign);
 }
 
-// For verbose BASH behaviour, use isatty(STDIN_FILENO); but exit() is enough
-// This doesn't check for overflow or long long datatypes
+static void	perror_numeric(t_shell *shell, const char *argv)
+{
+	ft_putstr_fd("minishell: exit: ", 2);
+	ft_putstr_fd((char *)argv, 2);
+	ft_putstr_fd(": numeric argument required\n", 2);
+	free_shell(shell);
+	exit(2);
+}
+
+/*	print "exit" before executing exit
+ 	handle: no args, 1 numeric, non-numeric & multiple
+	set status variable for multiple args (no exit)
+
+       	status = (unsigned char)status;
+		exit code valid range [0, 255] unsigned 8-bit
+*/
+int	ft_exit(t_shell *shell, char *argv[])
+{
+	long long	status;
+
+	status = 0;
+	write(1, "exit\n", 5);
+	if (!argv[1])
+		status = shell->last_status;
+	else if (!is_numeric(argv[1]))
+		perror_numeric(shell, argv[1]);
+	else if (argv[2])
+	{
+		ft_putstr_fd("minishell: exit: too many arguments\n", 2);
+		shell->last_status = 1;
+		return (1);
+	}
+	else
+		status = ft_strtoll(argv[1]);
+	status = (unsigned char)status;
+	rl_clear_history();
+	free_shell(shell);
+	exit(status);
+}
